Fix: save categories

+++ src/main/java/com/planify/backend/infrastructure/repositories/UsersRepository.java
@@ -8,10 +8,8 @@ import org.springframework.data.r2dbc.repository.R2dbcRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends R2dbcRepository<UsersEntity, Integer> {   
-    // Buscar por Google ID, ya no se usa, movido a AuthMethods
-    //Mono<UsersEntity> findByGoogleId(String googleId);
- 
-    // Buscar por email (útil en caso de verificaciones adicionales)
+public interface UsersRepository extends R2dbcRepository<UsersEntity, Long> {
+
     Mono<UsersEntity> findByEmail(String email);
+
 }
